#brisanje_predmeta Dodat dialog za potvrdu

+++ src/view/MenuBar.java
@@ -15,6 +15,7 @@ import com.sun.glass.events.KeyEvent;
 import controller.PredmetController;
 import controller.ProfesorController;
 import controller.StudentController;
+import model.PredmetBaza;
 import model.StudentBaza;
 
 
@@ -155,7 +156,17 @@ public class MenuBar extends JMenuBar {
 					break;
 				case 2:
 					int predmetRow = MainFrame.getInstance().getSelectedPredmetRow();
-					PredmetController.getInstance().deletePredmet(predmetRow);
+					if (predmetRow < 0) {
+						JOptionPane.showMessageDialog(MainFrame.getInstance(), "Izaberite predmet za brisanje!");
+						break;
+					}
+					
+					String predmetMessage = String.format("Predmet %s ce biti uklonjen iz baze. Nastaviti?", PredmetBaza.getInstance().getValueAt(predmetRow, 1));
+					int predmetOpt = JOptionPane.showConfirmDialog(MainFrame.getInstance(), predmetMessage, "Brisanje", JOptionPane.YES_NO_OPTION , JOptionPane.WARNING_MESSAGE);
+					
+					if(predmetOpt == 0)
+						PredmetController.getInstance().deletePredmet(predmetRow);
+					
 					break;
 				}
 			}
